Add testing.go with exported test helpers
- TestAccountDetails
- FetchTestAccountDetails
