+++ helper_test.go
@@ -27,32 +27,9 @@ func testClient(t *testing.T) *Client {
 	return client
 }
 
-// TestAccountDetails represents the info fetched from
-// /api/v2/account/details so we can determine the user
-// that is running the tests from the token being used.
-type TestAccountDetails struct {
-	ID       string `json:"id" jsonapi:"primary,users"`
-	Username string `jsonapi:"attr,username"`
-	Email    string `jsonapi:"attr,email"`
-}
-
-// fetchTestAccountDetails returns the username and external ID
-// of the user running the test
 func fetchTestAccountDetails(t *testing.T, client *Client) *TestAccountDetails {
 	if _testAccountDetails == nil {
-		tad := &TestAccountDetails{}
-		req, err := client.newRequest("GET", "account/details", nil)
-		if err != nil {
-			t.Fatalf("could not create request for user details: %v", err)
-		}
-
-		ctx := context.Background()
-		err = client.do(ctx, req, tad)
-		if err != nil {
-			t.Fatalf("could not fetch test user details: %v", err)
-		}
-		_testAccountDetails = tad
-		return tad
+		_testAccountDetails = FetchTestAccountDetails(t, client)
 	}
 	return _testAccountDetails
 }
